# docs/03_openapi_contract_draft.yaml
openapi: 3.1.0
info:
  title: Encrypted Data Query & Management API
  version: 0.1.0-draft
  description: |
    基于 SRS v1.0 的接口合同草案。
    所有接口默认返回 application/json，除文件上传外。
  contact:
    name: Project Team
servers:
  - url: /
    description: Local deployment
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: 用户列表（仅超级管理员）
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Users]
      summary: 创建用户（仅超级管理员）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/users/{userId}:
    patch:
      tags: [Users]
      summary: 更新用户角色或状态（仅超级管理员）
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: 删除用户（仅超级管理员）
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/users/{userId}/quota:
    put:
      tags: [Users]
      summary: 设置普通用户查询配额（仅超级管理员）
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuotaUpdateRequest'
      responses:
        '200':
          description: 配额更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidParam:
              value:
                code: INVALID_ARGUMENT
                message: 请求参数不合法
                trace_id: 67b8d10c5f1948f8
...
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                code: FORBIDDEN
                message: 无权限访问该资源
                trace_id: 3a2ee53fabff4d56
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                code: NOT_FOUND
                message: 资源不存在
                trace_id: 6f460f54b6f84607
  schemas:
    Role:
      type: string
      enum: [SUPER_ADMIN, ADMIN, USER]
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: success
    UserSummary:
      type: object
      required: [id, username, role, is_active]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        is_active:
          type: boolean
    UserResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/UserSummary'
...
    UserListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
    CreateUserRequest:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
        password:
          type: string
          minLength: 8
          maxLength: 128
        role:
          $ref: '#/components/schemas/Role'
    UpdateUserRequest:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/Role'
        is_active:
          type: boolean
    Quota:
      type: object
      properties:
        daily_limit:
          type: integer
        daily_used:
          type: integer
        total_limit:
          type: integer
        total_used:
          type: integer
    QuotaUpdateRequest:
      type: object
      required: [daily_limit, total_limit]
      properties:
        daily_limit:
          type: integer
          minimum: 0
        total_limit:
          type: integer
          minimum: 0
    QuotaResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Quota'
...
    ErrorResponse:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: string
          enum:
            - INVALID_ARGUMENT
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - QUOTA_EXCEEDED
            - IMPORT_FAILED
            - INTERNAL_ERROR
        message:
          type: string
        trace_id:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true